f1demo/backend/requirements.txt
fastapi
uvicorn[standard]
fastf1
httpx
schedule
flask
flask-cors
requests
feedparser
slowapi
redis
python-dateutil
